app/requirements.txt
bcrypt==4.1.2
cryptography==42.0.1
fastapi==0.103.0
Jinja2==3.1.2
python-jose==3.3.0
SQLAlchemy==2.0.15
uvicorn==0.22.0
websockets==12.0
